All of target "Perl Module"'s self tests are passing now except threading.
The pointers test was failing because the bridge bailed out on undef arguments when it should have passed NULL.

The exceptions test used an unhandled selector to test the bridge's ObjC->Perl exception-handling; checking for an invalid selector before making the call resulted in the test passing, but I don't think exceptions are really working properly yet.

The "BundleLoader" test had two problems: First, it tried to load "InterfaceBuilder.framework", which is no longer in a standard location as of Xcode 2.5. And second, the "CBDummyModule.pm" wasn't copied into the CB framework.